requirements.txt
google-cloud-storage
kafka-python
requests
psycopg2
pandas
